# internal/api/consent/v2/swagger.yml
openapi: 3.0.0
info:
  title: API Consents - Open Insurance Brasil
  description: |
    API que trata da criação, consulta e revogação de consentimentos para o Open Insurance Brasil - Fase 2 e 3.
    Não possui segregação entre pessoa natural e pessoa jurídica.

    # Orientações importantes
    A API Consents trata dos consentimentos para a fase 2 e 3 do Open Insurance Brasil.
    - As informações da instituição receptora não trafegam na API Consents – a autenticação da receptora se dá através do [DCR](https://br-openinsurance.github.io/areadesenvolvedor/#dynamic-client-registration-dcr).
    - Na chamada para a criação do consentimento deve-se utilizar um token gerado via `client_credentials`.
    - Após o `POST` de criação do consentimento, o `STATUS` devolvido na resposta deverá ser `AWAITING_AUTHORISATION`.
    - O `STATUS` será alterado para `AUTHORISED` somente após autenticação e confirmação por parte do usuário na instituição transmissora dos dados.
    - O `STATUS` será alterado para 'CONSUMED' após ocorrer o processamento da iniciação, seja ele com sucesso (HTTP 201) ou ainda em casos de insucesso (HTTP 422) retornados pela Transmissora. Para os demais códigos HTTP não há mudança de status do consentimento, o mesmo permanecerá AUTHORISED, respeitando o tempo máximo de expiração do consentimento.
    - Todas as datas trafegadas nesta API seguem o padrão da [RFC3339](https://tools.ietf.org/html/rfc3339) e formato "zulu".
...
     
...
     
     
      ```
...

...
      ```
...
  
...
servers:
...
paths:
  /consents:
    post:
      tags:
        - Consents
      summary: Criar novo pedido de consentimento.
      operationId: consentsPostConsents
      description: Método para a criação de um novo consentimento.
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: "#/components/parameters/x-v"
        - $ref: "#/components/parameters/x-min-v"
        - $ref: "#/components/parameters/xIdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsent'
        description: Payload para criação do consentimento.
        required: true
      responses:
        '201':
          $ref: '#/components/responses/201ConsentsCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '422':
          $ref: '#/components/responses/UnprocessableEntityConsents'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/Default'
      security:
        - OAuth2Security:
            - consents
  /consents/{consentId}:
    get:
      tags:
        - Consents
      summary: Obter detalhes do consentimento identificado por consentId.
      operationId: consentsGetConsentsConsentId
      description: Método para obter detalhes do consentimento identificado por consentId.
      parameters:
        - $ref: '#/components/parameters/consentId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: "#/components/parameters/x-v"
        - $ref: "#/components/parameters/x-min-v"
      responses:
        '200':
          $ref: '#/components/responses/200ConsentsConsentIdRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/Default'
      security:
        - OAuth2Security:
            - consents
    delete:
      tags:
        - Consents
      summary: Deletar / Revogar o consentimento identificado por consentId.
      operationId: consentsDeleteConsentsConsentId
      description: Método para deletar / revogar o consentimento identificado por consentId.
      parameters:
        - $ref: '#/components/parameters/consentId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/xFapiAuthDate'
        - $ref: '#/components/parameters/xFapiCustomerIpAddress'
        - $ref: '#/components/parameters/xFapiInteractionId'
        - $ref: '#/components/parameters/xCustomerUserAgent'
        - $ref: "#/components/parameters/x-v"
        - $ref: "#/components/parameters/x-min-v"
      responses:
        '204':
          $ref: '#/components/responses/204ConsentsConsentIdDeleted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/Default'
      security:
        - OAuth2Security:
            - consents
components:
  schemas:
...
    ResponseError:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          maxItems: 13
          items:
            type: object
            required:
              - code
              - title
              - detail
              - requestDateTime
            properties:
              code:
                description: Código de erro específico do endpoint
                type: string
                pattern: '[\w\W\s]*'
                maxLength: 255
              title:
                description: Título legível por humanos deste erro específico
                type: string
                pattern: '[\w\W\s]*'
                maxLength: 255
              detail:
                description: Descrição legível por humanos deste erro específico
                type: string
                pattern: '[\w\W\s]*'
                maxLength: 2048
              requestDateTime:
                description: Data e hora da consulta, conforme especificação RFC-3339, formato UTC.
                type: string
                maxLength: 20
                format: date-time
                example: '2021-05-21T08:30:00Z'
            additionalProperties: false
        meta:
          $ref: '#/components/schemas/Meta'
      additionalProperties: false
    422ResponseErrorCreateConsents:
      type: object
      required:
        - errors
      properties:
        errors:
          type: object
          minItems: 1
          required:
            - code
            - title
            - detail
          properties:
            code:
              type: string
              enum:
                - ERRO_IDEMPOTENCIA
                - NAO_INFORMADO
              example: ERRO_IDEMPOTENCIA
              description: Código do erro 422 de Entidade não processada.
            title:
              type: string
              maxLength: 255
              pattern: '[\w\W\s]*'
              example: Tentativa de alteração de requisição já processada
              description: |
                - ERRO_ IDEMPOTENCIA: Tentativa de alteração de requisição já processada
                - NÃO_INFORMADO: Não informada pelo servidor
            detail:
              type: string
              maxLength: 2048
              pattern: '[\w\W\s]*'
              example: Tentativa de alteração de requisição já processada
              description: |
                - ERRO_ IDEMPOTENCIA: Tentativa de alteração de requisição já processada
                - NÃO_INFORMADO: Não informada pelo servidor 
...
    XV:
        type: string
        description: |
          Versão do endpoint da API requisitado pelo cliente. O titular dos dados deve 
          responder com a versão mais alta suportada entre x-min-v e x-v. Se o valor de 
          x-min-v for igual ou maior que o valor de x-v, o cabeçalho x-min-v deve ser 
          tratado como ausente. Se todas as versões solicitadas não forem suportadas, 
          o titular dos dados deve responder com o código de status 406 Not Acceptable.
  parameters:
    consentId:
      name: consentId
      in: path
      description: |
        O consentId é o identificador único do consentimento a ser revogado e deverá ser um URN - Uniform Resource Name.
      required: true
      schema:
        type: string
        pattern: ^urn:[a-zA-Z0-9][a-zA-Z0-9-]{0,31}:[a-zA-Z0-9()+,\-.:=@;$_!*'%\/?#]+$
        maxLength: 256
    Authorization:
      name: Authorization
      in: header
      description: Cabeçalho HTTP padrão. Permite que as credenciais sejam fornecidas dependendo do tipo de recurso solicitado
      required: true
      schema:
        type: string
        pattern: '[\w\W\s]*'
        maxLength: 2048
    xCustomerUserAgent:
      name: x-customer-user-agent
      in: header
      description: Indica o user-agent que o usuário utiliza.
      required: false
      schema:
        type: string
        pattern: '[\w\W\s]*'
        minLength: 1
        maxLength: 100
    xFapiAuthDate:
      name: x-fapi-auth-date
      in: header
      description: 'Data em que o usuário logou pela última vez com o receptor. Representada de acordo com a [RFC7231](https://tools.ietf.org/html/rfc7231).Exemplo: Sun, 10 Sep 2017 19:43:31 UTC'
      required: false
      schema:
        type: string
        pattern: ^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec) \d{4} \d{2}:\d{2}:\d{2} (GMT|UTC)$
        minLength: 29
        maxLength: 29
    xFapiCustomerIpAddress:
      name: x-fapi-customer-ip-address
      in: header
      description: O endereço IP do usuário se estiver atualmente logado com o receptor.
      required: false
      schema:
        type: string
        pattern: '[\w\W\s]*'
        minLength: 1
        maxLength: 100
    xFapiInteractionId:
      name: x-fapi-interaction-id
      in: header
      description: 'Um UID [RFC4122](https://tools.ietf.org/html/rfc4122) usado como um ID de correlação. Se fornecido, o transmissor deve "reproduzir" esse valor no cabeçalho de resposta.'
      required: true
      schema:
        type: string
        pattern: ^[a-zA-Z0-9][a-zA-Z0-9\-]{0,99}$
        minLength: 1
        maxLength: 100
    x-v:
      name: x-v
      in: header
      description: |
        Versão do endpoint da API requisitado pelo cliente. O titular dos dados deve 
        responder com a versão mais alta suportada entre x-min-v e x-v. Se o valor de 
        x-min-v for igual ou maior que o valor de x-v, o cabeçalho x-min-v deve ser 
        tratado como ausente. Se todas as versões solicitadas não forem suportadas, 
        o titular dos dados deve responder com o código de status 406 Not Acceptable.
      required: false
      schema:
        type: string
      example: '2.1.3'
    x-min-v:
      name: x-min-v
      in: header
      description: |
        Versão mínima do endpoint da API requisitado pelo cliente. O detentor dos dados 
        deve responder com a versão mais alta suportada entre x-min-v e x-v. Se todas as 
        versões solicitadas não forem suportadas, o titular dos dados deve responder com 
        um código de status 406 Not Acceptable.
      required: false
      schema:
        type: string
      example: '2.0.0'
    xIdempotencyKey:
      name: x-idempotency-key
      in: header
      description: | 
        Cabeçalho HTTP personalizado. Identificador de solicitação 
        exclusivo para suportar a idempotência.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 40
        pattern: ^(?!\s)(.*)(\S)$
  securitySchemes: 
    OAuth2Security:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://authserver.example/token
          scopes:
            consents: Criação do consentimento.
  responses:
    UnprocessableEntityConsents:
      description: Seguir as orientações presentes na descrição deste endpoint
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/422ResponseErrorCreateConsents'
    UnprocessableEntity:
      description: O servidor entende o tipo de conteúdo da entidade da requisição, e a sintaxe da requisição esta correta, mas não foi possível processar as instruções presentes
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    BadRequest:
      description: A requisição foi malformada, omitindo atributos obrigatórios, seja no payload ou através de atributos na URL.
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    Forbidden:
      description: O token tem escopo incorreto ou uma política de segurança foi violada
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    InternalServerError:
      description: Ocorreu um erro no gateway da API ou no microsserviço
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    Default:
      description: Erro inesperado.
      headers:  
        x-v: 
          schema:
            $ref: '#/components/schemas/XV'
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    MethodNotAllowed:
      description: O consumidor tentou acessar o recurso com um método não suportado
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    NotAcceptable:
      description: A solicitação continha um cabeçalho Accept diferente dos tipos de mídia permitidos ou um conjunto de caracteres diferente de UTF-8
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    NotFound:
      description: O recurso solicitado não existe ou não foi implementado
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    TooManyRequests:
      description: A operação foi recusada, pois muitas solicitações foram feitas dentro de um determinado período ou o limite global de requisições concorrentes foi atingido
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    Unauthorized:
      description: Cabeçalho de autenticação ausente/inválido ou token inválido
      content:
        application/json; charset=utf-8:
          schema:
            $ref: '#/components/schemas/ResponseError'
    201ConsentsCreated:
      description: Consentimento criado com sucesso.
      headers:
        x-fapi-interaction-id:
          description: |
            Um UID [RFC4122](https://tools.ietf.org/html/rfc4122) usado como um ID de correlação.
            Se fornecido, o transmissor deve "reproduzir" esse valor no cabeçalho de resposta.
          schema:
            type: string
            pattern: ^[a-zA-Z0-9][a-zA-Z0-9\-]{0,99}$
            minLength: 1
            maxLength: 100
            example: 73cac523-d3ae-2289-b106-330a6218710d
        x-v: 
          schema:
            $ref: '#/components/schemas/XV'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseConsent'
    200ConsentsConsentIdRead:
      description: Consentimento consultado com sucesso.
      headers:
        x-fapi-interaction-id:
          description: |
            Um UID [RFC4122](https://tools.ietf.org/html/rfc4122) usado como um ID de correlação.
            Se fornecido, o transmissor deve "reproduzir" esse valor no cabeçalho de resposta.
          schema:
            type: string
            pattern: ^[a-zA-Z0-9][a-zA-Z0-9\-]{0,99}$
            minLength: 1
            maxLength: 100
            example: 92cac523-d3ae-2289-b106-330a6218710d
        x-v: 
          schema:
            $ref: '#/components/schemas/XV'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseConsent'
    204ConsentsConsentIdDeleted:
      description: Consentimento revogado com sucesso.
      headers:
        x-fapi-interaction-id:
          description: |
            Um UID [RFC4122](https://tools.ietf.org/html/rfc4122) usado como um ID de correlação.
            Se fornecido, o transmissor deve "reproduzir" esse valor no cabeçalho de resposta.
          schema:
            type: string
            pattern: ^[a-zA-Z0-9][a-zA-Z0-9\-]{0,99}$
            minLength: 1
            maxLength: 100
            example: 85bac523-d3ae-2289-b106-330a6218710d
        x-v: 
          schema:
            $ref: '#/components/schemas/XV'      